adding windows instructions and some small refactor as per reviews
